♻️ Extract gitmoji list in Home page
